move AdminNotifierCheckbox to its own file in groups
